Sourcery refactored master branch

+++ userbot/plugins/pm_logger.py
@@ -67,12 +67,11 @@ async def set_no_log_p_m(event):
     if Config.PM_LOGGR_BOT_API_ID is not None:
         event.pattern_match.group(1)
         chat = await event.get_chat()
-        if event.is_private:
-            if chat.id in NO_PM_LOG_USERS:
-                NO_PM_LOG_USERS.remove(chat.id)
-                await event.edit("Will Log Messages from this chat")
-                await asyncio.sleep(3)
-                await event.delete()
+        if event.is_private and chat.id in NO_PM_LOG_USERS:
+            NO_PM_LOG_USERS.remove(chat.id)
+            await event.edit("Will Log Messages from this chat")
+            await asyncio.sleep(3)
+            await event.delete()
 
 
 @bot.on(admin_cmd(pattern="nlog ?(.*)"))
@@ -80,12 +79,11 @@ async def set_no_log_p_m(event):
     if Config.PM_LOGGR_BOT_API_ID is not None:
         event.pattern_match.group(1)
         chat = await event.get_chat()
-        if event.is_private:
-            if chat.id not in NO_PM_LOG_USERS:
-                NO_PM_LOG_USERS.append(chat.id)
-                await event.edit("Won't Log Messages from this chat")
-                await asyncio.sleep(3)
-                await event.delete()
+        if event.is_private and chat.id not in NO_PM_LOG_USERS:
+            NO_PM_LOG_USERS.append(chat.id)
+            await event.edit("Won't Log Messages from this chat")
+            await asyncio.sleep(3)
+            await event.delete()
 
 CmdHelp("pm_logger").add_command(
   "save", "<reply>", "Saves the replied message to your pm logger group/channel"
